requirements.txt
astroid==2.13.3
asttokens==2.0.5
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
debugpy==1.6.3
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
exceptiongroup==1.1.0
executing==0.8.3
Flask==2.2.2
iniconfig==2.0.0
ipdb==0.13.11
ipykernel==6.16.1
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.3
jupyter_core==4.11.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib-inline==0.1.3
mccabe==0.7.0
mysql-connector-python==8.0.31
mysql
nest-asyncio==1.5.6
numpy
packaging==21.3
pandas
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.3
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.12.0
pylint==2.15.10
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.5
pyzmq==24.0.1
six==1.16.0
SQLAlchemy==2.0.1
stack-data==0.2.0
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.2.0
typing_extensions==4.4.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1